requirements.txt
# MindScope AI - Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Machine Learning
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
datasets==2.14.0
accelerate==0.24.0
captum==0.6.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.0

# Visualization & Explainability
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Report Generation
reportlab==4.0.7
jinja2==3.1.2

# Database & Storage
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.2
tqdm==4.66.1
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
